// src/shared/types/generated-types/paid-amounts-entity.ts
/**
 * Расчетный листок
 * API Расчетный листок
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { type PaidAmountEntity } from './paid-amount-entity';

export interface PaidAmountsEntity {
    paidAmounts: PaidAmountEntity[];
}