// src/shared/types/generated-types/payment-schedule-entity.ts
/**
 * Расчетный листок
 * API Расчетный листок
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { type PaymentScheduleItemEntity } from './payment-schedule-item-entity';

export interface PaymentScheduleEntity {
    paymentSchedule: PaymentScheduleItemEntity[];
}